+++ android/src/main/java/com/instabug/reactlibrary/RNInstabugReactnativeModule.java
@@ -9,6 +9,7 @@ import android.view.View;
 
 import com.facebook.react.bridge.Arguments;
 import com.facebook.react.bridge.Callback;
+import com.facebook.react.bridge.Promise;
 import com.facebook.react.bridge.ReactApplicationContext;
 import com.facebook.react.bridge.ReactMethod;
 import com.facebook.react.bridge.ReadableArray;
@@ -289,22 +290,20 @@ public class RNInstabugReactnativeModule extends EventEmitterModule {
      * @see #resetTags()
      */
     @ReactMethod
-    public void getTags(final Callback tagsCallback) {
+    public void getTags(final Promise promise) {
         MainThreadHandler.runOnMainThread(new Runnable() {
             @Override
             public void run() {
-                WritableArray tagsArray;
+                WritableArray tagsArray = Arguments.createArray();
                 try {
                     ArrayList<String> tags = Instabug.getTags();
-                    tagsArray = Arguments.createArray();
                     for (int i = 0; i < tags.size(); i++) {
                         tagsArray.pushString(tags.get(i));
                     }
-                    tagsCallback.invoke(tagsArray);
-
                 } catch (Exception e) {
                     e.printStackTrace();
                 }
+                promise.resolve(tagsArray);
             }
         });
     }
@@ -481,17 +480,17 @@ public class RNInstabugReactnativeModule extends EventEmitterModule {
      * @return the desired user attribute
      */
     @ReactMethod
-    public void getUserAttribute(final String key, final Callback userAttributeCallback) {
+    public void getUserAttribute(final String key, final Promise promise) {
         MainThreadHandler.runOnMainThread(new Runnable() {
             @Override
             public void run() {
-                String userAttribute;
+                String userAttribute = "";
                 try {
                     userAttribute = Instabug.getUserAttribute(key);
-                    userAttributeCallback.invoke(userAttribute);
                 } catch (Exception e) {
                     e.printStackTrace();
                 }
+                promise.resolve(userAttribute);
             }
         });
     }
@@ -522,7 +521,7 @@ public class RNInstabugReactnativeModule extends EventEmitterModule {
      * @return all user attributes as HashMap<String, String>
      */
     @ReactMethod
-    public void getAllUserAttributes(final Callback userAttributesCallback) {
+    public void getAllUserAttributes(final Promise promise) {
         MainThreadHandler.runOnMainThread(new Runnable() {
             @Override
             public void run() {
@@ -535,7 +534,7 @@ public class RNInstabugReactnativeModule extends EventEmitterModule {
                 } catch (Exception e) {
                     e.printStackTrace();
                 }
-                userAttributesCallback.invoke(writableMap);
+                promise.resolve(writableMap);
             }
         });
     }