+++ test/mocks/mockInstabugUtils.ts
@@ -1,5 +1,8 @@
 jest.mock('../../src/utils/InstabugUtils', () => {
+  const actual = jest.requireActual('../../src/utils/InstabugUtils');
+
   return {
+    ...actual,
     parseErrorStack: jest.fn(),
     captureJsErrors: jest.fn(),
     getActiveRouteName: jest.fn(),